remake isAddressSet, create role enum
