+++ src/main/java/com/example/demo/mapper/AddressMapper.java
@@ -13,7 +13,8 @@ import java.util.List;
 public interface AddressMapper {
     AddressResponseDto toDto(Address address);
 
-    void updateAddressFromDto(AddressRequestDto addressRequestDto,@MappingTarget Address address);
+    void updateAddressFromDto(AddressRequestDto addressRequestDto, @MappingTarget Address
+            address);
 
     List<AddressResponseDto> toDto(List<Address> address);
