+++ src/main/java/com/example/demo/service/impl/OrderItemServiceImpl.java
@@ -10,7 +10,6 @@ import com.example.demo.repositories.OrderRepository;
 import com.example.demo.repositories.ProductRepository;
 import com.example.demo.service.OrderItemService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,7 +19,6 @@ import static com.example.demo.util.ExceptionSourceName.PRODUCT;
 
 import java.util.List;
 
-@Slf4j
 @Service
 @RequiredArgsConstructor
 public class OrderItemServiceImpl implements OrderItemService {
@@ -33,7 +31,6 @@ public class OrderItemServiceImpl implements OrderItemService {
     @Override
     @Transactional
     public void addOrderItem(int orderId, OrderItemRequestDto orderItemRequestDto) {
-        log.info("Adding order item to order with id {}", orderId);
         OrderItem orderItem = orderItemMapper.toEntity(orderItemRequestDto);
         orderItem.setOrder(findOrderById(orderId));
         orderItem.setProduct(findProductById(orderItemRequestDto));
@@ -43,16 +40,14 @@ public class OrderItemServiceImpl implements OrderItemService {
     @Override
     @Transactional(readOnly = true)
     public List<OrderItemResponseDto> getAllOrderItems() {
-        log.info("Retrieving all order items");
         return orderItemMapper.toDto(orderItemRepository.findAll());
     }
 
     @Override
     @Transactional
     public OrderItemResponseDto updateOrderItem(int id, OrderItemRequestDto orderItemRequestDto) {
-        log.info("Updating order item with id {}", id);
         OrderItem orderItem = findOrderItemById(id);
-        orderItemMapper.updateOrderItemFromDto(orderItemRequestDto, orderItem);
+        orderItemMapper.updateOrderItemFromDto(orderItemRequestDto,orderItem);
         orderItem.setProduct(findProductById(orderItemRequestDto));
         orderItemRepository.save(orderItem);
         return orderItemMapper.toDto(orderItem);
@@ -61,33 +56,23 @@ public class OrderItemServiceImpl implements OrderItemService {
     @Override
     @Transactional
     public void deleteOrderItem(int id) {
-        log.info("Deleting order item with id {}", id);
         OrderItem orderItem = findOrderItemById(id);
         orderItemRepository.delete(orderItem);
     }
 
 
-    private OrderItem findOrderItemById(int id) {
+    private OrderItem findOrderItemById(int id){
         return orderItemRepository.findById(id)
-                .orElseThrow(() -> {
-                    log.error("Order item with id {} was not found", id);
-                    return new EntityNotFoundException(ORDER_ITEM, id);
-                });
+                .orElseThrow(() -> new EntityNotFoundException(ORDER_ITEM,id));
     }
 
-    private Product findProductById(OrderItemRequestDto orderItemRequestDto) {
+    private Product findProductById(OrderItemRequestDto orderItemRequestDto){
         return productRepository.findById(orderItemRequestDto.productId())
-                .orElseThrow(() -> {
-                    log.error("Product with id {} was not found", orderItemRequestDto.productId());
-                    return new EntityNotFoundException(PRODUCT, orderItemRequestDto.productId());
-                });
+                .orElseThrow(() -> new EntityNotFoundException(PRODUCT,orderItemRequestDto.productId()));
     }
 
-    private Order findOrderById(int id) {
+    private Order findOrderById(int id){
         return orderRepository.findById(id)
-                .orElseThrow(() -> {
-                    log.error("Order with id {} was not found", id);
-                    return new EntityNotFoundException(ORDER, id);
-                });
+                .orElseThrow(() -> new EntityNotFoundException(ORDER,id));
     }
 }
